Move non-String uninterruptible utilities to guest staging
Keep the arithmetic, atomic, primitive, and code helpers nested in UninterruptibleUtils so the migration preserves the original utility organization instead of introducing migration-only top-level classes.

Leave the substitution-dependent String families in core. Represent the two compiler bit operations as guest native markers and replace them with their existing Graal nodes through required hosted invocation plugins.

Add an end-to-end Native Image test that keeps both marker operands nonconstant, proves that no native call survives compilation, and verifies the generated operations against the JDK results.
