+++ substratevm/src/com.oracle.svm.guest.staging/src/com/oracle/svm/guest/staging/core/jdk/UninterruptibleUtils.java
@@ -24,6 +24,8 @@
  */
 package com.oracle.svm.guest.staging.core.jdk;
 
+import static com.oracle.svm.shared.Uninterruptible.CALLED_FROM_UNINTERRUPTIBLE_CODE;
+
 import org.graalvm.word.PointerBase;
 import org.graalvm.word.UnsignedWord;
 import org.graalvm.word.WordBase;
@@ -35,16 +37,19 @@ import com.oracle.svm.shared.util.VMError;
 import jdk.internal.misc.Unsafe;
 
 /**
- * Atomic values whose operations can be called from uninterruptible code. The implementations
- * inline the corresponding operations because the JDK atomic classes cannot be annotated with
- * {@link Uninterruptible}.
+ * Runtime helpers that can be called from uninterruptible code without depending on builder-owned
+ * compiler classes.
+ *
+ * Simple JDK-derived operations use local implementations annotated with {@link Uninterruptible}.
+ * The atomic helpers directly expose the required {@link Unsafe} operations because the JDK atomic
+ * classes cannot be annotated with {@link Uninterruptible}. Operations that need Graal compiler
+ * nodes are declared as native compiler markers and replaced by required hosted invocation plugins,
+ * so no native call remains in generated code.
+ *
+ * The String-related families remain temporarily in the core copy because they depend on the
+ * {@code Target_java_lang_String*} substitutions. Moving them is tracked by GR-77810.
  */
-public final class UninterruptibleAtomicUtils {
-
-    /** Prevents instantiation. */
-    private UninterruptibleAtomicUtils() {
-    }
-
+public class UninterruptibleUtils {
     public static class AtomicBoolean {
 
         private static final Unsafe UNSAFE = Unsafe.getUnsafe();
@@ -398,6 +403,151 @@ public final class UninterruptibleAtomicUtils {
         @SuppressWarnings("unchecked")
         public final T getAndSet(T newValue) {
             return (T) UNSAFE.getAndSetReference(this, VALUE_OFFSET, newValue);
+        }
+    }
+
+    /** Methods like the ones from {@link java.lang.Math} but annotated as uninterruptible. */
+    public static class Math {
+
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static int min(int a, int b) {
+            return (a <= b) ? a : b;
+        }
+
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static long min(long a, long b) {
+            return (a <= b) ? a : b;
+        }
+
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static int max(int a, int b) {
+            return (a >= b) ? a : b;
+        }
+
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static long max(long a, long b) {
+            return (a >= b) ? a : b;
+        }
+
+        @Uninterruptible(reason = CALLED_FROM_UNINTERRUPTIBLE_CODE, mayBeInlined = true)
+        public static double max(double a, double b) {
+            if (a != a) {
+                return a; // a is NaN
+            }
+            if ((a == 0.0d) && (b == 0.0d) && (Double.doubleToRawLongBits(a) == Double.doubleToRawLongBits(-0.0d))) {
+                // Raw conversion ok since NaN can't map to -0.0.
+                return b;
+            }
+            return (a >= b) ? a : b;
+        }
+
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static int clamp(int value, int min, int max) {
+            assert min <= max;
+            return min(max(value, min), max);
+        }
+
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static long clamp(long value, long min, long max) {
+            assert min <= max;
+            return min(max(value, min), max);
+        }
+
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static int abs(int a) {
+            return (a < 0) ? -a : a;
+        }
+
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static long abs(long a) {
+            return (a < 0) ? -a : a;
+        }
+
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static long floorToLong(double value) {
+            assert value == value : "must not be NaN";
+            return (long) value;
+        }
+
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static long ceilToLong(double a) {
+            long floor = floorToLong(a);
+            return a > floor ? floor + 1 : floor;
+        }
+    }
+
+    public static class Byte {
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        @SuppressWarnings("cast")
+        public static int toUnsignedInt(byte x) {
+            return ((int) x) & 0xff;
+        }
+    }
+
+    public static class Long {
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static native int countTrailingZeros(long i);
+
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static int hashCode(long value) {
+            return (int) (value ^ (value >>> 32));
+        }
+    }
+
+    public static class Integer {
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static int highestOneBit(int i) {
+            return i & (java.lang.Integer.MIN_VALUE >>> numberOfLeadingZeros(i));
+        }
+
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static native int numberOfLeadingZeros(int i);
+
+        /** Uninterruptible version of {@link java.lang.Integer#compare(int, int)}. */
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static int compare(int x, int y) {
+            return (x < y) ? -1 : ((x == y) ? 0 : 1);
+        }
+
+        /** Uninterruptible version of {@link java.lang.Integer#compareUnsigned(int, int)}. */
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static int compareUnsigned(int x, int y) {
+            return compare(x + java.lang.Integer.MIN_VALUE, y + java.lang.Integer.MIN_VALUE);
+        }
+    }
+
+    public static class CodeUtil {
+        @Uninterruptible(reason = CALLED_FROM_UNINTERRUPTIBLE_CODE, mayBeInlined = true)
+        public static long signExtend(long value, int inputBits) {
+            if (inputBits < 64) {
+                if ((value >>> (inputBits - 1) & 1) == 1) {
+                    return value | (-1L << inputBits);
+                } else {
+                    return value & ~(-1L << inputBits);
+                }
+            } else {
+                return value;
+            }
+        }
+
+        @Uninterruptible(reason = CALLED_FROM_UNINTERRUPTIBLE_CODE, mayBeInlined = true)
+        public static long zeroExtend(long value, int inputBits) {
+            if (inputBits < 64) {
+                return value & ~(-1L << inputBits);
+            } else {
+                return value;
+            }
+        }
+
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static boolean isPowerOf2(int val) {
+            return val > 0 && (val & val - 1) == 0;
+        }
+
+        @Uninterruptible(reason = "Called from uninterruptible code.", mayBeInlined = true)
+        public static int log2(int val) {
+            assert val > 0;
+            return (java.lang.Integer.SIZE - 1) - Integer.numberOfLeadingZeros(val);
         }
     }
 }
